requirements.txt
aiogram==2.9.2
emoji==0.5.4
moviepy==1.0.3
pytube3==9.6.4